Add files via upload
Sorting. Visualizer is a web app for visualizing a bunch of different sorting algorithms Like Selection Sort, Bubble Sort, Insertion Sort, Merge Sort, Quick Sort With the functionality of (Speed Control) and Array Size Control.

This project is built using HTML, CSS, and JS. This project sorting visualizer is a very simple UI and it allows the users to select the sort algorithm, select the array size, and speed of the visualization. Also there is a stop button which will help you to stop the sorting of the array elements.